+++ CHANGELOG.md
@@ -1,4 +1,26 @@
-## Unreleased - TBD
+## TBD
+
+### Bug Fixes
+
+* Fix [[#146](https://github.com/dbcli/litecli/issues/146)], making sure `.once`
+  can be used more than once in a session.
+
+## 1.9.0 - 2022-06-06
+
+### Features
+
+* Add support for ANSI escape sequences for coloring the prompt.
+* Add support for `.indexes` command.
+* Add an option to turn off the auto-completion menu. Completion menu can be
+  triggered by pressed the `<tab>` key when this option is set to False. Fixes
+  [#105](https://github.com/dbcli/litecli/issues/105).
+
+### Bug Fixes
+
+* Fix [#120](https://github.com/dbcli/litecli/issues/120). Make the `.read` command actually read and execute the commands from a file.
+* Fix  [#96](https://github.com/dbcli/litecli/issues/96) the crash in VI mode when pressing `r`.
+
+## 1.8.0 - 2022-03-29
 
 ### Features
 
@@ -9,6 +31,7 @@
 ### Bug Fixes
 
 * Upgrade cli_helpers to workaround Pygments regression.
+* Use get_terminal_size from shutil instead of click. 
 
 ## 1.7.0 - 2022-01-11